fixed typecast int to float

+++ VGAnuS_Software-Ontwikkeling/EE-API-library/Src/API_draw.c
@@ -20,7 +20,7 @@ int API_draw_line(uint16_t x, uint8_t y, uint16_t x2, uint8_t y2, uint8_t color,
 	uint8_t err = 0;
 	float a;
 	float b;
-	float f;
+	int f;
 
 	a = (y2 - y) / (x2 - x);
 	b = y - x*a;
